requirements.txt
pillow  12.2.0
pip     26.0.1